requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
black==21.9b0
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-cors-headers==3.9.0
django-environ==0.7.0
djangorestframework==3.12.4
environ==1.0
flake8==3.9.2
iniconfig==1.1.1
install==1.3.4
isort==5.9.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.1
regex==2021.9.24
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
